Quark: cmactions prevent crash of Irgestures sensor
On Quark, if you contantly enable and disable the irgestures after
some attempts (is random) it will get "lost" and will not work
properly after until a reboot, this can happen in hours or days.

Prevent from ever disable it after hand wave is enable
and only send a sensor action if the screen is off.

In the past was not intended to disable the irsensor as it
"appears to be buggy" originally described on
Iff32deff4490d3aeb4aabbc2a05afbee6bead140

Change-Id: I02366649fd856bb0bd03d25eaca74848b09da

Signed-off-by: Felipe Leon <fglfgl27@gmail.com>